inserção de autorização e autenticação em todos os endpoints permitindo apenas selecionar,modificar,deletar 'todos' do proprio usuario

+++ ToDoApp/routers/todos.py
@@ -27,19 +27,23 @@ class TodoRequest(BaseModel):
   complete: bool
 
 @router.get("/",status_code= status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-  todos =  db.query(Todos).all()
+async def read_all(user:user_dependency,db: db_dependency):
+  if user is None:
+    raise HTTPException(status_code=401, detail='Authetication Failed')
+  #retornar todos os todos com base no usuário
+  todos =  db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
   return todos
 
 @router.get("/todo/{todo_id}", status_code= status.HTTP_200_OK)
-async def read_todo(db : db_dependency, todo_id : int = Path (gt=0)):
-  
-
-  todo_model =  db.query(Todos).filter(Todos.id == todo_id).first() #chave unica, logo nao precisa percorrer banco inteiro para achar outros ids 
+async def read_todo(user:user_dependency,db : db_dependency, todo_id : int = Path (gt=0)):
+  if user is None:
+    raise HTTPException(status_code=401, detail='Authetication Failed')
+  #retorna todos com base no id, porém não quero que qualquer usuário veja todas os todos 
+  todo_model =  db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).first() #chave unica, logo nao precisa percorrer banco inteiro para achar outros ids 
   if todo_model is not None:
     return todo_model
   
-  raise HTTPException ( status_code = 404, detail ='Todo Not found') 
+  raise HTTPException (status_code = 404, detail ='Todo Not found') 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
 async def create_todo(user: user_dependency,db: db_dependency, todo_request: TodoRequest):
@@ -50,26 +54,34 @@ async def create_todo(user: user_dependency,db: db_dependency, todo_request: Tod
   db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,
+async def update_todo(user:user_dependency,db:db_dependency,
                        todo_request : TodoRequest,
                       todo_id: int = Path(gt=0)
                      ):
-  todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+  if user is None:
+    raise HTTPException(status_code=401, detail='Authetication Failed')
+  
+  todo_model = db.query(Todos).filter(Todos.id == todo_id,Todos.owner_id == user.get('id')).first() 
   if todo_model is None:
     raise HTTPException(status_code=404, detail = 'Todo not found')
     
   todo_model.title = todo_request.title
   todo_model.description = todo_request.description 
-  todo_model.priority = todo_request.complete
+  todo_model.priority = todo_request.priority
   todo_model.complete = todo_request.complete
+
   db.add(todo_model)
   db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo (db:db_dependency,
+async def delete_todo (user:user_dependency,db:db_dependency,
                        todo_id: int =  Path(gt = 0)):
-  todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+  if user is None:
+    raise HTTPException(status_code=401, detail='Authetication Failed')
+
+  todo_model = db.query(Todos).filter(Todos.id == todo_id,Todos.owner_id == user.get('id')).first()
   if todo_model is None:
     raise HTTPException(status_code=404, detail = 'Todo not found')
-  db.query(Todos).filter(Todos.id == todo_id).delete()
+  db.query(Todos).filter(Todos.id == todo_id,Todos.owner_id == user.get('id')).delete()
+  
   db.commit()
